Fix updating the database after app launch
There is a bug in Android (or maybe only in OnePlus) that cause the
system usage events to be removed after system update. In previous
implementation the app would first remove logs from the previous week
and then populate the database with new values. Right now, the app
first checks if there are any logs before removing previous values.