Fix crashed that occurred after app opening
When opening the app, TodaysStatsFragment is the first fragment that
appears on the screen. It has a method for loading the ad after 500
milliseconds and then adding the ad to the adapter. But sometimes user
navigated back from that fragment causing the adapter to be
uninitialized again and therefore when the ad finished loading it
threw an UninitializedPropertyAccessException when trying to access
that adapter variable.