Wrap await() function in a try catch block
Without this, an exception would be thrown when a user navigates away
from the app before it finishes loading the data. ViewModel would call
dispose() inside the onCleared() that would interrupt the thread waiting
for the data loading process to finish. The interruption would cause
a crash because InterrputedException would have nowhere to go.